fix: :bug: Mejoras
Cambiar de int a float en montos, Cancelar, Ver Stock siendo un socio

+++ src/Model/Table/SecondaryProductsTable.php
@@ -64,7 +64,7 @@ class SecondaryProductsTable extends Table
             ->allowEmptyString('id', null, 'create');
 
         $validator
-            ->integer('stock')
+            ->decimal('stock')
             ->notEmptyString('stock');
 
         $validator
